[DO NOT MERGE] Quadplane parametric model (#23)
* Added DynamicsModel class as base for all generated dynamics models"
The model class contains properties shared between all models and shgall simplyfy automated checks and the later
export to a sitl gazebo model by providing a unified interface for all models.
Furthermore, the existing simple_multirotor model was adapted to use the new class.

* Adapted check for all required topics and added method to compute resampled dataframe

* started fixing rotor parameters

* initial model structure for quad plane

* Transformed rotor model into class
the simple multirotor model was transformed into a class to make it easier to build up upon it in other models.

* fixed typo and worked on Quad Plane Model Structure

* added data imports and possibility to run the quadplane model from main.py

* Adapted dynamics Model class and the two current model to enable renaming the ulog data.
Since the naming inside the ulog file is often not very understandable and sometimes contains the same name for different physical properties (e.g. xyz for both angular velocity and angular accelerations) there is a need to being able to rename the columns of the dataframe. Otherwise we might end up with multiple columns with the same name.
This was implemented in a way that renaming can be done if needed and the two present models were adapted.

* added input normalization and computation of airspeed in quad plane model

* added symmetric logistic sigmoid function for lift model and deleted no longer needed runge-kutta file

* Feature Matrix for regression of actuator Thrust forces implemented"

* fixed normalization of input 4 and typo

* added output vector

* added Linear+Plate Aerodynamics Model

* added linear + plate aero model

* added lin+plate  aero model to quad_plane_model

* added regression

* added transformation of world acceleration and fixed symmetric logistic sigmoid finction

* added plot of predicted accelerations

* moved coordinate transformation to method of dynamics model

* added actuator normalization to dynamics_model

* added more sufisticated dram model

* added airspeed plots

* implemented actuator force model as it is used in gazebo_motor_model.cpp

* fixed wrong naming in gazebo rotor model

* new rotor model added to quad_plane_model

* added side force due to different displacement of the two ailerons

* added slerp interpolation to dataframe resampling

* cleaned up structure for model plots

* added aero model description and systematic coefficient list

* added improved terminal output and model param to yaml file

* adapted rotor model