Quad plane improvements and Cleanup of YAML Result Files (#36)
* Safeguard for actuator normalization and fixing stall angle conversion
A safeguard ensuring that the actuator normalization never produces negative values was added. This was needed since for example vertical rotors of the standart vtol quadplane have actuator values of 900 in fixed wing mode, even though their minimal pwm is specified as 1000.

Secondly the stall angle was currently specified in degrees without converting it to radians. This conversion is needed to use it together with the angle of attack and the trigonometric functions of the math library. The default stall angle was increased to 35 degrees for now.

* Added folder model structure for test logs
I decided to move the each model to it's own folder inside models. This way we can store all test logs, tests and performance results of this model inside its corresponding directory.

* Improved Model structure and removed tracking of yaml result files

* Added log file to output yaml

* Added test logs and test results

Ensured that inflow scalar is positive semidefinite

Remove overriding PX4_ROOT


Fix log path and result paths

* Remove result files from source code

* Ignore yaml files

* Add smaller logs

Co-authored-by: Jaeyoung-Lim <jalim@ethz.ch>