First pytest to check transformation A first pytest was added to check the integrity of the transformation methods of the DynamicsModel. Hereby it it load attitude quaternions and acceleration data in both inertial frame and (acceleration - gravity vector) data as measured by the accelerometer in body frame. For this reason also the gravity vector is substracted from the acceleration in inertial frame to achieve comparibal data.
In a next step the two transforms (NED inertial to FRD body frame and vise versa) are applied and the RMSE between the transformed and measured acceleration is computed. The RMSE was observed to be in both cases around 0.078 m/s^2. Therefore, the test for now just thresholds both rmse to be less than 0.1m/s^2.