Adapting rotor model to estimate inflow scaler Instead of hardcoding the an inflow scaler we now estimate the airspeed, perpendicular to the rotor axis, at which the propeller does not produce any thrust anymore.
In short: The blades of the propeller expierence a relative airflow composed of the inflow of the airspeed of the vehicle and the rotation of the blade itself. At a certain angle of attack of the airflow relative to the blade the propeller blade will not produce any lift anymore and sebsequentially the rotor does not procude any thrust anymore. Since the ratio between incoming airspeed and blade speed due to the angular velocity is responsible for this angle of attack we can derive a model for rotor thrust depending on angular velocity and angular velocity squared. More details will follow in the model description pdf.