Fixed actuator sequence and position Fixed the sequence and position of the actuators to be consistent with the standart_vtol_aerodynamics.sdf file. This, thogether with the turning direction is needed to generate a meaningfull output of the moments in the rotor model.
Currently, when using the gazebo_rotor_model we are required to have the position and turning direction of each rotor. I think this is generally a good way since those inputs should be fairly easy to obtain on a drone and simplify the validation and comparison of the generated model.