Aeromodel structure rework This commit includes multiple smaller changes to the structure of the aero models:
First and foremost, the aero model was renamed to be consistent with the PX4 convention. I.E. AAE stands for aileron, aileron, elevator.

The present aero model has been adapted to use the cropped symmetric sigmoid function to solve a problem described in #62.

Furthermore, the scale factor for the symmetric sigmoid function is now contained in the config file.

Last but not least I fixed a naming mistake where coefficients from the forces in xz direction would be named with "xy".