Unified Rotor feature calculation and separate moment and forces estimation
In this commit the calculation of the rotor features was moved to a method in the dynamics model and is therefore now unified between all the models. The user can configure rotor groups in the config file of his model. Hereby, all rotors from the same group are assumed to be physically (motor, prop, voltage, ect.) equal and will share all coefficients as determined by the parametric model.

Furthermore, i decided to allow the user to configure whether forces, moments or both are estimated by the algorithm. This was combined with the changes mentioned above since the rotor force and moment features share a significant part of the computation. In particular both need to compute the airplow parallel and perpendicular to the rotor axis for each timestamp. With the current implementation those computations are shared when both are activated which results in much faster performance for large datasets.