Adding additional rotor models Commit adapts the standard rotor model to specify the rotor axis (direction of thrust force) for each computation of the force or moment feature matrix. This is needed to make the model combatiple with with its childs. In the standard model always the fix rotor axis is used for this computation.
The changing axis rotor model is a child of the rotor model and introduces a matrix of rotor axis where each row represents the axis for the corresponding timestep. This model itself is not intended to be in a model, as can be seen by the fact, that it just constructs the matrix by all identical rotor axis vectors. Instead it is ment as a common basis for the other tho new models.

The bi-directional rotor model is a child of the changing axis rotor model. Hereby the rotor can turn in both directions which subsequentially results in the thrust force changing direction. The model handels this, by expecting a normalized actuatir input vector between [-1, 1]. Hereby, the sign of the entry is used to switch the force directions for the negative inputs.

Last but not least a Tilting Rotor model is introduced. This model can be used to model rotors, where the direction of the rotoraxis can be adjusted by a separate actuator as can be done in tiltwings or tiltrotor models. Also this model is a child of the changing axis model and overrides the function for the creation of the rotor axis matrix.