Dig model config file paths out of source
The assumption that the config file is under a specific relative path is only valid for the default config files. This makes it harder to handle cases where the configuration files are in other places.

This PR removes the default config file path from the source, and handles it by using default config file names that are the same as the model name. The user can always override the config file by speficying the path to the config file