Conditional package import of Visual Dataframe Selector
This commit modifies the package import behaviour to only import the visual dataframe selector when the corresponding argument is set to enable it. This way the toolchain can be run without installing the additional dependencies for the VDS if the operator wishes to do so. It also eliminates some errors that came up when pyqt tried to set up its GUI connections when run in the containers used for the SysID CI.