Update README
This commit removes the instructions to run the pipeline directly with python. I think specifying and supporting multiple versions of running the program is potentially just confusing for new operators while providing minimal additional value.

Furthermore, The commit extended on the documentation of the different arguments of the `estimate-model` command.