Generalizing and unifying model class naming
This commit renames the quadrotor model class to the multirotor model class. The quadrotor class was already capable of using different rotor configuration. This way the class can be generalized.

To stay consisten with our default arguments for configuration file and default log, the model argument passed when using `estimate-model` is still the quadrotor. This makes sense since config and log file are specific to the quadrotor.

Furthermore it made the naming of the other classes a bit more consistent and renamed all config files to now be consistent with the model names.