Adapting Naming of quadplane accelerations
I already started to adapt the naming of the accelerations to reflect the data property (e.g. acceleration in body frame in x direction) instead of the name of the source topic. This way the properties can be accesed in the code independent from which source topic the operator chooses.