Adding covariance heatmap
The covariance heatmap is added as a measure to asses how well coefficients are correlated in the underlying data. Hereby, a high correlation means that the effect of the coefficients on the output (allecerlation prediciton) is very similar and that the two coefficients can not be observed separatly from each other.